Moved internal xaml to a new internal namespace
Made sure modalitylayout does not do anything if we try to hide it when it is not hidden
