+++ src/DIPS.Xamarin.UI/DIPS.Xamarin.UI.csproj
@@ -41,13 +41,13 @@
   </ItemGroup>
 
   <ItemGroup>
-    <EmbeddedResource Update="Controls\RadioButtonGroup\RadioButton.xaml">
+    <EmbeddedResource Update="Internal\Xaml\RadioButton.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
     <EmbeddedResource Update="Controls\RadioButtonGroup\RadioButtonGroup.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
-    <EmbeddedResource Update="Controls\Sheet\SheetView.xaml">
+    <EmbeddedResource Update="Internal\Xaml\SheetView.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
     <EmbeddedResource Update="Resources\LocalizedStrings\InternalLocalizedStrings.resx">
