+++ src/DIPS.Xamarin.UI/Internal/Xaml/RadioButton.xaml
@@ -5,7 +5,7 @@
              xmlns:d="http://xamarin.com/schemas/2014/forms/design"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              mc:Ignorable="d"
-             x:Class="DIPS.Xamarin.UI.Controls.RadioButtonGroup.RadioButton"
+             x:Class="DIPS.Xamarin.UI.Internal.xaml.RadioButton"
              x:Name="codeBehind">
     <ContentView.GestureRecognizers>
         <TapGestureRecognizer Tapped="OnTapped" />
