+++ src/Tests/DIPS.Xamarin.UI.Tests/Converters/ValueConverters/DateAndTimeConverterTests.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Globalization;
 using DIPS.Xamarin.UI.Converters.ValueConverters;
-using DIPS.Xamarin.UI.InternalUtils;
+using DIPS.Xamarin.UI.Internal.Utilities;
 using DIPS.Xamarin.UI.Resources.LocalizedStrings;
 using DIPS.Xamarin.UI.Tests.TestHelpers;
 using FluentAssertions;
